Added check for wrong deleted marker

+++ lost_link/local/local_file_processor.py
@@ -1,3 +1,4 @@
+import os
 from langchain_chroma import Chroma
 
 from lost_link.ai.embedding_generator import EmbeddingGenerator
@@ -22,6 +23,11 @@ class LocalFileProcessor:
         self._embedding_generator.generate_and_store_embeddings(file.path, LocalFile, file.id)
 
     def _process_deleted_file(self, file: LocalFile):
+        if os.path.exists(file.path):
+            # Datei existiert noch, wurde fälschlicherweise als deleted markiert, also als bearbeitet behandeln
+            file.deleted = False
+            self._process_edited_file(file)
+            return
         ids_to_delete = [x.id for x in file.embeddings]
         if len(ids_to_delete) > 0:
             self._vector_db.delete(ids_to_delete)
@@ -50,8 +56,8 @@ class LocalFileProcessor:
         for file in self._file_manager.get_all_edited_files():
             self._process_edited_file(file)
 
-        for file in self._file_manager.get_all_new_files():
-            self._process_new_file(file)
-
         for file in self._file_manager.get_all_deleted_files():
             self._process_deleted_file(file)
+
+        for file in self._file_manager.get_all_new_files():
+            self._process_new_file(file)
